adds C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -9,13 +9,13 @@ int main(void)
 {
 char char1;
 int int1;
-int int2;
+long int int2;
 long long int longlongint1;
 float float1;
 
 printf("Size of a char: %ld bytes(s) \n", sizeof(char1));
 printf("Size of an int: %ld byte(s)\n", sizeof(int1));
-printf("Size of a long int: %ld byte(s)\n", sizeof(int2));
+printf("Size of a long int: %li byte(s)\n", sizeof(int2));
 printf("Size of a long long int: %ld byte(s)\n", sizeof(longlongint1));
 printf("Size of a float: %ld byte(s)\n", sizeof(float1));
 return (0);
